chore(refactor): moved config into graph component; some other small chores
